Factor boilerplate into kaleido_main.cc:
And make it part of the dependency.
