+++ src/kaleido/cc/126.0.6478.126/kaleido.cc
@@ -6,66 +6,11 @@
 
 #include "headless/app/kaleido.h"
 
-// Fundamental chromium includes
-#include "content/public/app/content_main.h"
-#include "headless/lib/headless_content_main_delegate.h"
-
-// Fundamental utilities
-#include "base/logging.h"
-#include "base/files/file_util.h"
-
-#include "build/build_config.h" // IS_WIN and stuff like that
-
-// Sandbox Includes
-#if BUILDFLAG(IS_WIN)
-#include "content/public/app/sandbox_helper_win.h"
-#include "sandbox/win/src/sandbox_types.h"  // nogncheck
-#elif BUILDFLAG(IS_MAC)
-#include "base/check.h"
-#include "sandbox/mac/seatbelt_exec.h"
-#endif
-
-// CLI includes, not sure if using all
-#include "base/command_line.h"
-#include "base/base_switches.h"
-#include "content/public/common/content_switches.h"
-#include "headless/public/switches.h"
-#if BUILDFLAG(IS_MAC)
-#include "components/os_crypt/sync/os_crypt_switches.h"  // nogncheck
-#endif
-
-// Browser Includes
 #include "headless/lib/browser/headless_browser_impl.h"
 #include "headless/public/headless_browser.h"
-
-#if BUILDFLAG(IS_WIN)
-#include "base/strings/utf_string_conversions.h"
-#include "components/crash/core/app/crash_switches.h"  // nogncheck
-#include "components/crash/core/app/run_as_crashpad_handler_win.h"
-#include "sandbox/win/src/sandbox_types.h"
-#endif
-
-// Can you clear up deps in build?
+#include "base/functional/bind.h"
 
 namespace kaleido {
-
-// Kaleido manages a browser and its tabs
-class Kaleido {
- public:
-  Kaleido() = default;
-
-  Kaleido(const Kaleido&) = delete;
-  Kaleido& operator=(const Kaleido&) = delete;
-
-  ~Kaleido() = default;
-
-  void OnBrowserStart(headless::HeadlessBrowser* browser);
-
- private:
-  void ShutdownSoon();
-  void Shutdown();
-  raw_ptr<headless::HeadlessBrowser> browser_ = nullptr;
-};
 
 void Kaleido::OnBrowserStart(headless::HeadlessBrowser* browser) {
   browser_ = browser;
@@ -82,180 +27,7 @@ void Kaleido::Shutdown() {
   browser_.ExtractAsDangling()->Shutdown();
 }
 
-//  Much of KaleidoMain is boilerplate taking from headless/app/ example:
-//  - it starts sandboxes, which may be pointless, but our flags are chaos
-//      - init_tools flags no sandbox
-//      - here we initialize it
-//      - python then turns it off again
-//      - it is not really necessary
-//  - it, depending on platform, moves argc and argv towards a HeadlessBrowser instance
-//
-//  It is better not to pass whatever chromium flag into kaleido,
-//  unless there was a flag specifically for that "--chromium_flags="--whatever=23,-f," etc
-//  This function will be called several times as several processes are started by the main process
-//  with a variety of command line flags.
-//  Filtering out flags may not be reasonable in this case, and some of the switches deleted may be necessary:
-//  HeadlessChildMain
-//  HEADLESS_USE_CRASHPAD -> kCrashpadHandler
-
-int KaleidoMain(int argc, const char** argv) {
-  content::ContentMainParams params(nullptr); // TODO  WHAT IS THIS REALLY FOR
-
-// LETS CONSTRUCT SANDBOX THAT WE THEN DISABLE
-#if BUILDFLAG(IS_WIN)
-  sandbox::SandboxInterfaceInfo sandbox_info = {nullptr};
-  content::InitializeSandboxInfo(&sandbox_info);
-  // Sandbox info has to be set and initialized.
-  params.sandbox_info = &sandbox_info;
-#if BUILDFLAG(IS_MAC)
-  sandbox::SeatbeltExecServer::CreateFromArgumentsResult seatbelt =
-      sandbox::SeatbeltExecServer::CreateFromArguments(
-          argv[0], argc, const_cast<char**>(argv));
-  if (seatbelt.sandbox_required) {
-    CHECK(seatbelt.server->InitializeSandbox());
-  }
-#endif  // BUILDFLAG(IS_MAC)
-#endif  // BUILDFLAG(IS_WIN)
-
-#if BUILDFLAG(IS_WIN)
-  base::CommandLine::Init(0, nullptr);
-  // It's a good way to process CommandLine, but is windows really not capable of using it?
-#else
-  base::CommandLine::Init(argc, argv);
-#endif  // BUILDFLAG(IS_WIN)
-
-  base::CommandLine& command_line(*base::CommandLine::ForCurrentProcess());
-  std::string process_type = command_line.GetSwitchValueASCII(::switches::kProcessType);
-  LOG(INFO) << "Process type: " << process_type;
-  // command_line.AppendSwitch(::switches::kDisableGpu); // <-- possibility
-
-#if defined(HEADLESS_USE_CRASHPAD)
-  if (process_type == crash_reporter::switches::kCrashpadHandler) {
-    return crash_reporter::RunAsCrashpadHandler(
-        *base::CommandLine::ForCurrentProcess(), base::FilePath(),
-        ::switches::kProcessType, switches::kUserDataDir);
-  }
-#endif  // defined(HEADLESS_USE_CRASHPAD)
-
-// BELOW IS A TEMPORARY MUST-REMOVE TEST
-#if BUILDFLAG(IS_WIN)
-#if defined(HEADLESS_USE_CRASHPAD)
-  LOG(FATAL) << "crashpad IS used on windows, reactivate comments. Need command_line boilerplate." << std::endl;
-#else
-  LOG(FATAL) << "we can get rid of all crashpad" << std::endl;
-#endif
-#endif
-  // Some Logging
-  LOG(INFO) << "Original command: " << command_line.GetArgumentsString();
-  LOG(INFO) << "Args size: " << command_line.GetArgs().size();
-  for (const auto &piece : command_line.GetArgs()) {
-    LOG(INFO) << piece << std::endl;
-  }
-  if (!process_type.empty()) {
-    headless::HeadlessContentMainDelegate delegate(nullptr);
-    params.delegate = &delegate;
-    int rc = content::ContentMain(std::move(params));
-    base::Process::TerminateCurrentProcessImmediately(rc);
-    NOTREACHED_IN_MIGRATION();
-  }
-  // So we must be the main process...
-
-// EXAMPLE SAYS WE (MAC USERS) NEED THIS
-#if BUILDFLAG(IS_MAC)
-  command_line.AppendSwitch(os_crypt::switches::kUseMockKeychain);
-#endif
-
-
-  // Now we're going to start the browser
-  Kaleido kmanager;
-  auto browser = std::make_unique<headless::HeadlessBrowserImpl>(
-      base::BindOnce(&Kaleido::OnBrowserStart, base::Unretained(&kmanager)));
-  headless::HeadlessContentMainDelegate delegate(std::move(browser));
-  params.delegate = &delegate;
-  return content::ContentMain(std::move(params));
-  // return EXIT_FAILURE; // save for future use, where does EXIT_FAILURE come from?
-}
 
 
 } // namespace kaleido
 
-/*
-#include <memory>
-
-#include "base/functional/bind.h"
-#include "base/task/thread_pool.h"
-#include "build/branding_buildflags.h"
-#include "build/build_config.h"
-#include "headless/lib/browser/headless_web_contents_impl.h"
-#include "headless/public/headless_browser_context.h"
-#include "headless/public/headless_web_contents.h"
-
-#include "net/base/filename_util.h"
-#include "url/gurl.h"
-
-
-
-namespace kaleido {
-
-namespace {
-
-#if BUILDFLAG(IS_WIN)
-const wchar_t kAboutBlank[] = L"about:blank";
-#else
-const char kAboutBlank[] = "about:blank";
-#endif
-
-
-};
-
-
-  HeadlessBrowserContext::Builder context_builder =
-      browser_->CreateBrowserContextBuilder();
-
-  // Create browser  context and set it as the default. The default browser
-  // context is used by the Target.createTarget() DevTools command when no other
-  // context is given.
-  HeadlessBrowserContext* browser_context = context_builder.Build();
-  browser_->SetDefaultBrowserContext(browser_context);
-
-  const bool devtools_enabled = static_cast<HeadlessBrowserImpl*>(browser)
-                                    ->options()
-                                    ->DevtoolsServerEnabled();
-
-  // If no explicit URL is present navigate to about:blank unless we're being
-  // driven by a debugger.
-  base::CommandLine::StringVector args =
-      base::CommandLine::ForCurrentProcess()->GetArgs();
-  args.erase(
-      std::remove(args.begin(), args.end(), base::CommandLine::StringType()),
-      args.end());
-
-  if (args.empty() && !devtools_enabled) {
-    args.push_back(kAboutBlank);
-  }
-
-  if (args.empty()) {
-    return;
-  }
-
-  GURL target_url = ConvertArgumentToURL(args.front());
-  HeadlessWebContents::Builder builder(
-      browser_context->CreateWebContentsBuilder());
-
-  // If driven by a debugger just open the target page and
-  // leave expecting the debugger will do what they need.
-  if (devtools_enabled) {
-    HeadlessWebContents* web_contents =
-        builder.SetInitialURL(target_url).Build();
-    if (!web_contents) {
-      LOG(ERROR) << "Navigation to " << target_url << " failed.";
-      ShutdownSoon();
-    }
-    return;
-  }
-
-
-}  // namespace
-
-}  // namespace headless
-*/
